update response middleware

+++ routes/localSubjectAccessReview.js
@@ -10,14 +10,14 @@ router.get(routes.map((e) => `${e}/:name`), validSchema(apiAppsV1OpenApiV3), gen
 
 router.get(['/api/v1/localsubjectaccessreviews', ...routes], validSchema(apiV1OpenApiV3), general.find(LocalSubjectAccessReview), general.format(LocalSubjectAccessReview), general.list(LocalSubjectAccessReview));
 
-router.post(routes, validSchema(apiAppsV1OpenApiV3), general.save(LocalSubjectAccessReview));
+router.post(routes, validSchema(apiAppsV1OpenApiV3), general.save(LocalSubjectAccessReview), general.sendObj(LocalSubjectAccessReview));
 
-router.put(routes, validSchema(apiAppsV1OpenApiV3), general.update(LocalSubjectAccessReview));
+router.put(routes, validSchema(apiAppsV1OpenApiV3), general.update(LocalSubjectAccessReview), general.sendObj(LocalSubjectAccessReview));
 
-router.patch(routes.map((e) => `${e}/:name`), validSchema(apiAppsV1OpenApiV3), general.patch(LocalSubjectAccessReview));
+router.patch(routes.map((e) => `${e}/:name`), validSchema(apiAppsV1OpenApiV3), general.patch(LocalSubjectAccessReview), general.sendObj(LocalSubjectAccessReview));
 
-router.delete(routes.map((e) => `${e}/:name`), validSchema(apiAppsV1OpenApiV3), general.deleteOne(LocalSubjectAccessReview));
+router.delete(routes.map((e) => `${e}/:name`), validSchema(apiAppsV1OpenApiV3), general.deleteOne(LocalSubjectAccessReview), general.sendObj(LocalSubjectAccessReview));
 
-router.delete(routes, validSchema(apiAppsV1OpenApiV3), general.delete(LocalSubjectAccessReview));
+router.delete(routes, validSchema(apiAppsV1OpenApiV3), general.delete(LocalSubjectAccessReview), general.sendObj(LocalSubjectAccessReview));
 
 module.exports = router;
